requirements.txt
django>=4.2
requests>=2.31.0
python-whois>=0.8.0
python-dotenv>=1.0.0
